+++ tests/wp-i18n-polyfill.php
@@ -22,6 +22,10 @@
 
 declare( strict_types = 1 );
 
+if ( ! defined( 'HOUR_IN_SECONDS' ) ) {
+	define( 'HOUR_IN_SECONDS', 60 * 60 );
+}
+
 if ( ! defined( 'DAY_IN_SECONDS' ) ) {
 	define( 'DAY_IN_SECONDS', 24 * 60 * 60 );
 }